[PRDP-183] Reformat BizEventToReceiptService
